restructuring architecture for scaling
